Move all classes under edu.hm.hafner.analysis.
