+++ src/main/java/edu/hm/hafner/analysis/parser/DrMemoryParser.java
@@ -14,7 +14,7 @@ import edu.hm.hafner.analysis.ParsingException;
 import edu.hm.hafner.analysis.Severity;
 import edu.hm.hafner.util.LookaheadStream;
 
-import static edu.hm.hafner.util.IntegerParser.*;
+import static edu.hm.hafner.analysis.util.IntegerParser.*;
 
 /**
  * A parser for the Dr. Memory Errors.
