+++ src/main/java/edu/hm/hafner/analysis/parser/IdeaInspectionParser.java
@@ -15,7 +15,7 @@ import edu.hm.hafner.analysis.ParsingException;
 import edu.hm.hafner.analysis.ReaderFactory;
 import edu.hm.hafner.analysis.Report;
 import edu.hm.hafner.analysis.Severity;
-import edu.hm.hafner.util.XmlElementUtil;
+import edu.hm.hafner.analysis.util.XmlElementUtil;
 
 /**
  * A parser for IntelliJ IDEA inspections.
