+++ src/main/java/edu/hm/hafner/analysis/parser/fxcop/FxCopParser.java
@@ -12,7 +12,7 @@ import edu.hm.hafner.analysis.ParsingCanceledException;
 import edu.hm.hafner.analysis.ParsingException;
 import edu.hm.hafner.analysis.ReaderFactory;
 import edu.hm.hafner.analysis.Report;
-import edu.hm.hafner.util.XmlElementUtil;
+import edu.hm.hafner.analysis.util.XmlElementUtil;
 
 /**
  * Parses a fxcop xml report file.
