+++ src/main/java/edu/hm/hafner/analysis/registry/PmdDescriptor.java
@@ -4,7 +4,7 @@ import edu.hm.hafner.analysis.Issue;
 import edu.hm.hafner.analysis.IssueParser;
 import edu.hm.hafner.analysis.parser.pmd.PmdMessages;
 import edu.hm.hafner.analysis.parser.pmd.PmdParser;
-import edu.hm.hafner.util.Deferred;
+import edu.hm.hafner.analysis.util.Deferred;
 
 /**
  * A descriptor for the Pmd warnings.
