+++ src/main/java/edu/hm/hafner/analysis/util/IntegerParser.java
@@ -1,4 +1,4 @@
-package edu.hm.hafner.util;
+package edu.hm.hafner.analysis.util;
 
 import org.apache.commons.lang3.StringUtils;
