+++ src/main/java/edu/hm/hafner/analysis/util/XmlElementUtil.java
@@ -1,4 +1,4 @@
-package edu.hm.hafner.util;
+package edu.hm.hafner.analysis.util;
 
 import java.util.ArrayList;
 import java.util.List;
