+++ src/main/java/edu/hm/hafner/analysis/util/package-info.java
@@ -5,7 +5,7 @@
  * @author Ullrich Hafner
  */
 @DefaultAnnotation(NonNull.class)
-package edu.hm.hafner.util;
+package edu.hm.hafner.analysis.util;
 
 import edu.umd.cs.findbugs.annotations.DefaultAnnotation;
 import edu.umd.cs.findbugs.annotations.NonNull;
